test: keep model discovery proof at integration boundary
