+++ src/agents/models-config.providers.auth-provenance.test.ts
@@ -758,79 +758,6 @@ describe("models-config provider auth provenance", () => {
     });
   });
 
-  it("publishes the canonically selected OpenAI account through implicit model config", async () => {
-    const profileA = "openai:profile-a";
-    const profileB = "openai:profile-b";
-    const keyA = "rejected-profile-a";
-    const keyB = "selected-profile-b";
-    const config: OpenClawConfig = {
-      auth: {
-        order: {
-          openai: [profileB, profileA],
-        },
-      },
-    };
-    const store: AuthProfileStore = {
-      version: 1,
-      profiles: {
-        [profileA]: { type: "api_key", provider: "openai", key: keyA },
-        [profileB]: { type: "api_key", provider: "openai", key: keyB },
-      },
-    };
-    const requests: string[] = [];
-    const outcomes: Array<import("../plugins/provider-catalog.types.js").ProviderCatalogOutcome> =
-      [];
-    const agentDir = tempDirs.make("openai-profile-order-");
-    const { buildOpenAIProvider } = await import("../../extensions/openai/api.js");
-    const { planOpenClawModelsJson } = await import("./models-config.plan.js");
-    discovery.providers = [buildOpenAIProvider()];
-    const fetchSpy = vi.spyOn(globalThis, "fetch").mockImplementation(async (_input, init) => {
-      const authorization = new Headers(init?.headers).get("authorization") ?? "";
-      requests.push(authorization);
-      if (authorization === `Bearer ${keyA}`) {
-        return new Response("unauthorized", { status: 401 });
-      }
-      if (authorization === `Bearer ${keyB}`) {
-        return Response.json({ data: [{ id: "gpt-5.5", object: "model" }] });
-      }
-      throw new Error("unexpected OpenAI catalog authorization");
-    });
-
-    try {
-      const plan = await planOpenClawModelsJson({
-        context: {
-          cfg: config,
-          discoveryAuthConfig: config,
-          sourceConfigForSecrets: config,
-          agentDir,
-          env: {},
-          envFingerprint: {},
-          providerDiscoveryProviderIds: ["openai"],
-          onProviderCatalogOutcome: (outcome) => outcomes.push(outcome),
-        },
-        authStore: store,
-        existingRaw: "",
-        existingParsed: null,
-      });
-
-      expect(requests).toEqual([`Bearer ${keyB}`]);
-      expect(outcomes).toEqual([{ provider: "openai", profileId: profileB, status: "ready" }]);
-      expect(plan.action).toBe("write");
-      const providers =
-        plan.action === "write"
-          ? (
-              JSON.parse(plan.contents) as {
-                providers?: Record<string, { models?: Array<{ id?: string }> }>;
-              }
-            ).providers
-          : undefined;
-      expect(providers?.openai?.models?.map((model) => model.id)).toContain("gpt-5.5");
-    } finally {
-      fetchSpy.mockRestore();
-      discovery.providers = [];
-    }
-  });
-
   it("resolves plugin-owned synthetic auth through the provider hook", () => {
     // Plugin-owned synthetic auth can provide discovery keys while persisted
     // config still records a non-secret marker.
